v 0.2.1 - minor update, wayne madden
fixed issues in bitmap causing last 8 bits to be unitinitialised and
last element of matrix to not be stored. nextFree calculation was using
itemsize instead of byte count. allowed for setting of existing element
in matrix. created cuda memory manager types which are used to access
matrix elements during kernal call, 'copy' should be called before
kernel is called.